+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 ## unreleased
 1. Fix json: cannot unmarshal object into Go struct field Message.message.parts of type types.Part (#5)
+2. Add APIKey, HTTP, OAuth2, and OpenIdConnect security schemes, structures and constants.